[issue-3] extract "fractals" into file, add testing
